+++ src/lib/cimb_translator/CimbDecoder.h
@@ -14,13 +14,13 @@ class CimbDecoder
 public:
 	CimbDecoder(unsigned symbol_bits, unsigned color_bits, bool dark=true, uchar ahashThreshold=0);
 
-	void update_color_correction(cv::Matx<double, 3, 3>&& ccm);
+	void update_color_correction(cv::Matx<float, 3, 3>&& ccm);
 
 	unsigned get_best_symbol(image_hash::ahash_result& results, unsigned& drift_offset, unsigned& best_distance) const;
 	unsigned decode_symbol(const cv::Mat& cell, unsigned& drift_offset, unsigned& best_distance) const;
 	unsigned decode_symbol(const bitmatrix& cell, unsigned& drift_offset, unsigned& best_distance) const;
 
-	unsigned get_best_color(double r, double g, double b) const;
+	unsigned get_best_color(float r, float g, float b) const;
 	unsigned decode_color(const Cell& cell) const;
 
 	bool expects_binary_threshold() const;
@@ -31,7 +31,7 @@ protected:
 	bool load_tiles();
 
 	unsigned check_color_distance(std::tuple<uchar,uchar,uchar> a, std::tuple<uchar,uchar,uchar> b) const;
-	std::tuple<uchar,uchar,uchar> fix_color(std::tuple<double,double,double> c, double adjustUp, double down) const;
+	std::tuple<uchar,uchar,uchar> fix_color(std::tuple<float,float,float> c, float adjustUp, float down) const;
 
 protected:
 	std::vector<uint64_t> _tileHashes;
